Extracted repetitive shader code into an abstract class. Also updated gitignore to exclude all shader header files.
