+++ Puzzle/Puzzle.vcxproj.filters
@@ -65,6 +65,9 @@
     <ClCompile Include="boardstateclass.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="shaderinterface.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="systemclass.h">
@@ -110,6 +113,9 @@
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="boardstateclass.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="shaderinterface.h">
       <Filter>Header Files</Filter>
     </ClInclude>
   </ItemGroup>
