+++ Puzzle/colorshaderclass.cpp
@@ -4,16 +4,12 @@
 #include "colorshaderclass.h"
 
 
-ColorShaderClass::ColorShaderClass()
+ColorShaderClass::ColorShaderClass() : ShaderInterface()
 {
-	m_vertexShader = nullptr;
-	m_pixelShader = nullptr;
-	m_layout = nullptr;
-	m_matrixBuffer = nullptr;
 }
 
 
-ColorShaderClass::ColorShaderClass(const ColorShaderClass& other)
+ColorShaderClass::ColorShaderClass(const ColorShaderClass& other) : ShaderInterface(other)
 {
 }
 
@@ -25,70 +21,10 @@ ColorShaderClass::~ColorShaderClass()
 
 bool ColorShaderClass::Initialize(ID3D11Device* device)
 {
-	bool result;
-
-
-	// Initialize the vertex and pixel shaders.
-	result = InitializeShader(device);
-	if (!result)
-	{
-		return false;
-	}
-
-	return true;
-}
-
-
-void ColorShaderClass::Shutdown()
-{
-	// Shutdown the vertex and pixel shaders as well as the related objects.
-	ShutdownShader();
-
-	return;
-}
-
-
-bool ColorShaderClass::Render(ID3D11DeviceContext* deviceContext, int indexCount, XMMATRIX worldMatrix, XMMATRIX viewMatrix,
-	XMMATRIX projectionMatrix)
-{
-	bool result;
-
-
-	// Set the shader parameters that it will use for rendering.
-	result = SetShaderParameters(deviceContext, worldMatrix, viewMatrix, projectionMatrix);
-	if (!result)
-	{
-		return false;
-	}
-
-	// Now render the prepared buffers with the shader.
-	RenderShader(deviceContext, indexCount);
-
-	return true;
-}
-
-
-bool ColorShaderClass::InitializeShader(ID3D11Device* device)
-{
-	HRESULT result;
 	D3D11_INPUT_ELEMENT_DESC polygonLayout[2];
 	unsigned int numElements;
-	D3D11_BUFFER_DESC matrixBufferDesc;
+	bool result;
 
-
-	// Create the vertex shader from the header file provided by compiling VertexShader.hlsl.
-	result = device->CreateVertexShader(g_colorvs, sizeof(g_colorvs), nullptr, &m_vertexShader);
-	if (FAILED(result))
-	{
-		return false;
-	}
-
-	// Create the pixel shader from the header file provided by compiling PixelShader.hlsl.
-	result = device->CreatePixelShader(g_colorps, sizeof(g_colorps), nullptr, &m_pixelShader);
-	if (FAILED(result))
-	{
-		return false;
-	}
 
 	// Create the vertex input layout description.
 	// This setup needs to match the VertexType stucture in the ModelClass and in the shader.
@@ -111,24 +47,9 @@ bool ColorShaderClass::InitializeShader(ID3D11Device* device)
 	// Get a count of the elements in the layout.
 	numElements = sizeof(polygonLayout) / sizeof(polygonLayout[0]);
 
-	// Create the vertex input layout.
-	result = device->CreateInputLayout(polygonLayout, numElements, g_colorvs, sizeof(g_colorvs), &m_layout);
-	if (FAILED(result))
-	{
-		return false;
-	}
-
-	// Setup the description of the dynamic matrix constant buffer that is in the vertex shader.
-	matrixBufferDesc.Usage =				D3D11_USAGE_DYNAMIC;
-	matrixBufferDesc.ByteWidth =			sizeof(MatrixBufferType);
-	matrixBufferDesc.BindFlags =			D3D11_BIND_CONSTANT_BUFFER;
-	matrixBufferDesc.CPUAccessFlags =		D3D11_CPU_ACCESS_WRITE;
-	matrixBufferDesc.MiscFlags =			0;
-	matrixBufferDesc.StructureByteStride =	0;
-
-	// Create the constant buffer pointer so we can access the vertex shader constant buffer from within this class.
-	result = device->CreateBuffer(&matrixBufferDesc, nullptr, &m_matrixBuffer);
-	if (FAILED(result))
+	// Initialize the vertex and pixel shaders.
+	result = InitializeShader(device, g_colorvs, sizeof(g_colorvs), g_colorps, sizeof(g_colorps), polygonLayout, numElements);
+	if (!result)
 	{
 		return false;
 	}
@@ -137,91 +58,17 @@ bool ColorShaderClass::InitializeShader(ID3D11Device* device)
 }
 
 
-void ColorShaderClass::ShutdownShader()
+void ColorShaderClass::Shutdown()
 {
-	// Release the matrix constant buffer.
-	if (m_matrixBuffer)
-	{
-		m_matrixBuffer->Release();
-		m_matrixBuffer = nullptr;
-	}
-
-	// Release the layout.
-	if (m_layout)
-	{
-		m_layout->Release();
-		m_layout = nullptr;
-	}
-
-	// Release the pixel shader.
-	if (m_pixelShader)
-	{
-		m_pixelShader->Release();
-		m_pixelShader = nullptr;
-	}
-
-	// Release the vertex shader.
-	if (m_vertexShader)
-	{
-		m_vertexShader->Release();
-		m_vertexShader = nullptr;
-	}
+	// Shutdown the vertex and pixel shaders as well as the related objects.
+	ShutdownShader();
 
 	return;
 }
 
 
-bool ColorShaderClass::SetShaderParameters(ID3D11DeviceContext* deviceContext, XMMATRIX worldMatrix, XMMATRIX viewMatrix,
-	XMMATRIX projectionMatrix)
+void ColorShaderClass::DrawShader(ID3D11DeviceContext* deviceContext, int indexCount, int)
 {
-	HRESULT result;
-	D3D11_MAPPED_SUBRESOURCE mappedResource;
-	MatrixBufferType* dataPtr;
-	unsigned int bufferNumber;
-
-
-	// Transpose the matrices to prepare them for the shader.
-	worldMatrix = XMMatrixTranspose(worldMatrix);
-	viewMatrix = XMMatrixTranspose(viewMatrix);
-	projectionMatrix = XMMatrixTranspose(projectionMatrix);
-
-	// Lock the constant buffer so it can be written to.
-	result = deviceContext->Map(m_matrixBuffer, 0, D3D11_MAP_WRITE_DISCARD, 0, &mappedResource);
-	if (FAILED(result))
-	{
-		return false;
-	}
-
-	// Get a pointer to the data in the constant buffer.
-	dataPtr = (MatrixBufferType*)mappedResource.pData;
-
-	// Copy the matrices into the constant buffer.
-	dataPtr->world = worldMatrix;
-	dataPtr->view = viewMatrix;
-	dataPtr->projection = projectionMatrix;
-
-	// Unlock the constant buffer.
-	deviceContext->Unmap(m_matrixBuffer, 0);
-
-	// Set the position of the constant buffer in the vertex shader.
-	bufferNumber = 0;
-
-	// Finally set the constant buffer in the vertex shader with the updated values.
-	deviceContext->VSSetConstantBuffers(bufferNumber, 1, &m_matrixBuffer);
-
-	return true;
-}
-
-
-void ColorShaderClass::RenderShader(ID3D11DeviceContext* deviceContext, int indexCount)
-{
-	// Set the vertex input layout.
-	deviceContext->IASetInputLayout(m_layout);
-
-	// Set the vertex and pixel shaders that will be used to render this triangle.
-	deviceContext->VSSetShader(m_vertexShader, nullptr, 0);
-	deviceContext->PSSetShader(m_pixelShader, nullptr, 0);
-
 	// Render the geometry.
 	deviceContext->DrawIndexed(indexCount, 0, 0);
