+++ Puzzle/colorshaderclass.h
@@ -4,17 +4,10 @@
 #pragma once
 
 
-//////////////
-// INCLUDES //
-//////////////
-#include <d3d11.h>
-#include <d3dcompiler.h>
-#include <directxmath.h>
-
-
 ///////////////////////
 // MY CLASS INCLUDES //
 ///////////////////////
+#include "shaderinterface.h"
 #include "color.ps.h"
 #include "color.vs.h"
 
@@ -23,35 +16,16 @@ using namespace DirectX;
 ////////////////////////////////////////////////////////////////////////////////
 // Class name: ColorShaderClass
 ////////////////////////////////////////////////////////////////////////////////
-class ColorShaderClass
+class ColorShaderClass : public ShaderInterface
 {
-private:
-	struct MatrixBufferType
-	{
-		XMMATRIX world;
-		XMMATRIX view;
-		XMMATRIX projection;
-	};
-
 public:
 	ColorShaderClass();
 	ColorShaderClass(const ColorShaderClass&);
 	~ColorShaderClass();
 
-	bool Initialize(ID3D11Device*);
-	void Shutdown();
-	bool Render(ID3D11DeviceContext*, int, XMMATRIX, XMMATRIX, XMMATRIX);
+	bool Initialize(ID3D11Device*) override;
+	void Shutdown() override;
 
-private:
-	bool InitializeShader(ID3D11Device*);
-	void ShutdownShader();
-
-	bool SetShaderParameters(ID3D11DeviceContext*, XMMATRIX, XMMATRIX, XMMATRIX);
-	void RenderShader(ID3D11DeviceContext*, int);
-
-private:
-	ID3D11VertexShader*	m_vertexShader;
-	ID3D11PixelShader*	m_pixelShader;
-	ID3D11InputLayout*	m_layout;
-	ID3D11Buffer*		m_matrixBuffer;
+protected:
+	void DrawShader(ID3D11DeviceContext*, int, int) override;
 };
