+++ Puzzle/instanceshaderclass.cpp
@@ -4,16 +4,12 @@
 #include "instanceshaderclass.h"
 
 
-InstanceShaderClass::InstanceShaderClass()
+InstanceShaderClass::InstanceShaderClass() : ShaderInterface()
 {
-	m_vertexShader = nullptr;
-	m_pixelShader = nullptr;
-	m_layout = nullptr;
-	m_matrixBuffer = nullptr;
 }
 
 
-InstanceShaderClass::InstanceShaderClass(const InstanceShaderClass& other)
+InstanceShaderClass::InstanceShaderClass(const InstanceShaderClass& other) : ShaderInterface(other)
 {
 }
 
@@ -25,11 +21,50 @@ InstanceShaderClass::~InstanceShaderClass()
 
 bool InstanceShaderClass::Initialize(ID3D11Device* device)
 {
+	D3D11_INPUT_ELEMENT_DESC polygonLayout[4];
+	unsigned int numElements;
 	bool result;
 
 
+	// Create the vertex input layout description.
+	// This setup needs to match the VertexType stucture in the ModelClass and in the shader.
+	polygonLayout[0].SemanticName =			"POSITION";
+	polygonLayout[0].SemanticIndex =		0;
+	polygonLayout[0].Format =				DXGI_FORMAT_R32G32B32_FLOAT;
+	polygonLayout[0].InputSlot =			0;
+	polygonLayout[0].AlignedByteOffset =	0;
+	polygonLayout[0].InputSlotClass =		D3D11_INPUT_PER_VERTEX_DATA;
+	polygonLayout[0].InstanceDataStepRate =	0;
+
+	polygonLayout[1].SemanticName =			"COLOR";
+	polygonLayout[1].SemanticIndex =		0;
+	polygonLayout[1].Format =				DXGI_FORMAT_R32G32B32A32_FLOAT;
+	polygonLayout[1].InputSlot =			0;
+	polygonLayout[1].AlignedByteOffset =	D3D11_APPEND_ALIGNED_ELEMENT;
+	polygonLayout[1].InputSlotClass =		D3D11_INPUT_PER_VERTEX_DATA;
+	polygonLayout[1].InstanceDataStepRate =	0;
+
+	polygonLayout[2].SemanticName =			"POSITION";
+	polygonLayout[2].SemanticIndex =		1;
+	polygonLayout[2].Format =				DXGI_FORMAT_R32G32B32_FLOAT;
+	polygonLayout[2].InputSlot =			1;
+	polygonLayout[2].AlignedByteOffset =	0;
+	polygonLayout[2].InputSlotClass =		D3D11_INPUT_PER_INSTANCE_DATA;
+	polygonLayout[2].InstanceDataStepRate =	1;
+
+	polygonLayout[3].SemanticName =			"COLOR";
+	polygonLayout[3].SemanticIndex =		1;
+	polygonLayout[3].Format =				DXGI_FORMAT_R32G32B32A32_FLOAT;
+	polygonLayout[3].InputSlot =			1;
+	polygonLayout[3].AlignedByteOffset =	D3D11_APPEND_ALIGNED_ELEMENT;
+	polygonLayout[3].InputSlotClass =		D3D11_INPUT_PER_INSTANCE_DATA;
+	polygonLayout[3].InstanceDataStepRate =	1;
+
+	// Get a count of the elements in the layout.
+	numElements = sizeof(polygonLayout) / sizeof(polygonLayout[0]);
+
 	// Initialize the vertex and pixel shaders.
-	result = InitializeShader(device);
+	result = InitializeShader(device, g_instancevs, sizeof(g_instancevs), g_instanceps, sizeof(g_instanceps), polygonLayout, numElements);
 	if (!result)
 	{
 		return false;
@@ -48,196 +83,8 @@ void InstanceShaderClass::Shutdown()
 }
 
 
-bool InstanceShaderClass::Render(ID3D11DeviceContext* deviceContext, int indexCountPerInstance, int instanceCount, XMMATRIX worldMatrix, XMMATRIX viewMatrix,
-	XMMATRIX projectionMatrix)
+void InstanceShaderClass::DrawShader(ID3D11DeviceContext* deviceContext, int indexCountPerInstance, int instanceCount)
 {
-	bool result;
-
-
-	// Set the shader parameters that it will use for rendering.
-	result = SetShaderParameters(deviceContext, worldMatrix, viewMatrix, projectionMatrix);
-	if (!result)
-	{
-		return false;
-	}
-
-	// Now render the prepared buffers with the shader.
-	RenderShader(deviceContext, indexCountPerInstance, instanceCount);
-
-	return true;
-}
-
-
-bool InstanceShaderClass::InitializeShader(ID3D11Device* device)
-{
-	HRESULT result;
-	D3D11_INPUT_ELEMENT_DESC polygonLayout[4];
-	unsigned int numElements;
-	D3D11_BUFFER_DESC matrixBufferDesc;
-
-
-	// Create the vertex shader from the header file provided by compiling VertexShader.hlsl.
-	result = device->CreateVertexShader(g_instancevs, sizeof(g_instancevs), nullptr, &m_vertexShader);
-	if (FAILED(result))
-	{
-		return false;
-	}
-
-	// Create the pixel shader from the header file provided by compiling PixelShader.hlsl.
-	result = device->CreatePixelShader(g_instanceps, sizeof(g_instanceps), nullptr, &m_pixelShader);
-	if (FAILED(result))
-	{
-		return false;
-	}
-
-	// Create the vertex input layout description.
-	// This setup needs to match the VertexType stucture in the ModelClass and in the shader.
-	polygonLayout[0].SemanticName =			"POSITION";
-	polygonLayout[0].SemanticIndex =		0;
-	polygonLayout[0].Format =				DXGI_FORMAT_R32G32B32_FLOAT;
-	polygonLayout[0].InputSlot =			0;
-	polygonLayout[0].AlignedByteOffset =	0;
-	polygonLayout[0].InputSlotClass =		D3D11_INPUT_PER_VERTEX_DATA;
-	polygonLayout[0].InstanceDataStepRate = 0;
-
-	polygonLayout[1].SemanticName =			"COLOR";
-	polygonLayout[1].SemanticIndex =		0;
-	polygonLayout[1].Format =				DXGI_FORMAT_R32G32B32A32_FLOAT;
-	polygonLayout[1].InputSlot =			0;
-	polygonLayout[1].AlignedByteOffset =	D3D11_APPEND_ALIGNED_ELEMENT;
-	polygonLayout[1].InputSlotClass =		D3D11_INPUT_PER_VERTEX_DATA;
-	polygonLayout[1].InstanceDataStepRate = 0;
-
-	polygonLayout[2].SemanticName =			"POSITION";
-	polygonLayout[2].SemanticIndex =		1;
-	polygonLayout[2].Format =				DXGI_FORMAT_R32G32B32_FLOAT;
-	polygonLayout[2].InputSlot =			1;
-	polygonLayout[2].AlignedByteOffset =	0;
-	polygonLayout[2].InputSlotClass =		D3D11_INPUT_PER_INSTANCE_DATA;
-	polygonLayout[2].InstanceDataStepRate = 1;
-
-	polygonLayout[3].SemanticName =			"COLOR";
-	polygonLayout[3].SemanticIndex =		1;
-	polygonLayout[3].Format =				DXGI_FORMAT_R32G32B32A32_FLOAT;
-	polygonLayout[3].InputSlot =			1;
-	polygonLayout[3].AlignedByteOffset =	D3D11_APPEND_ALIGNED_ELEMENT;
-	polygonLayout[3].InputSlotClass =		D3D11_INPUT_PER_INSTANCE_DATA;
-	polygonLayout[3].InstanceDataStepRate = 1;
-
-	// Get a count of the elements in the layout.
-	numElements = sizeof(polygonLayout) / sizeof(polygonLayout[0]);
-
-	// Create the vertex input layout.
-	result = device->CreateInputLayout(polygonLayout, numElements, g_instancevs, sizeof(g_instancevs), &m_layout);
-	if (FAILED(result))
-	{
-		return false;
-	}
-
-	// Setup the description of the dynamic matrix constant buffer that is in the vertex shader.
-	matrixBufferDesc.Usage =			   D3D11_USAGE_DYNAMIC;
-	matrixBufferDesc.ByteWidth =		   sizeof(MatrixBufferType);
-	matrixBufferDesc.BindFlags =		   D3D11_BIND_CONSTANT_BUFFER;
-	matrixBufferDesc.CPUAccessFlags =	   D3D11_CPU_ACCESS_WRITE;
-	matrixBufferDesc.MiscFlags =		   0;
-	matrixBufferDesc.StructureByteStride = 0;
-
-	// Create the constant buffer pointer so we can access the vertex shader constant buffer from within this class.
-	result = device->CreateBuffer(&matrixBufferDesc, nullptr, &m_matrixBuffer);
-	if (FAILED(result))
-	{
-		return false;
-	}
-
-	return true;
-}
-
-
-void InstanceShaderClass::ShutdownShader()
-{
-	// Release the matrix constant buffer.
-	if (m_matrixBuffer)
-	{
-		m_matrixBuffer->Release();
-		m_matrixBuffer = nullptr;
-	}
-
-	// Release the layout.
-	if (m_layout)
-	{
-		m_layout->Release();
-		m_layout = nullptr;
-	}
-
-	// Release the pixel shader.
-	if (m_pixelShader)
-	{
-		m_pixelShader->Release();
-		m_pixelShader = nullptr;
-	}
-
-	// Release the vertex shader.
-	if (m_vertexShader)
-	{
-		m_vertexShader->Release();
-		m_vertexShader = nullptr;
-	}
-
-	return;
-}
-
-
-bool InstanceShaderClass::SetShaderParameters(ID3D11DeviceContext* deviceContext, XMMATRIX worldMatrix, XMMATRIX viewMatrix,
-	XMMATRIX projectionMatrix)
-{
-	HRESULT result;
-	D3D11_MAPPED_SUBRESOURCE mappedResource;
-	MatrixBufferType* dataPtr;
-	unsigned int bufferNumber;
-
-
-	// Transpose the matrices to prepare them for the shader.
-	worldMatrix = XMMatrixTranspose(worldMatrix);
-	viewMatrix = XMMatrixTranspose(viewMatrix);
-	projectionMatrix = XMMatrixTranspose(projectionMatrix);
-
-	// Lock the constant buffer so it can be written to.
-	result = deviceContext->Map(m_matrixBuffer, 0, D3D11_MAP_WRITE_DISCARD, 0, &mappedResource);
-	if (FAILED(result))
-	{
-		return false;
-	}
-
-	// Get a pointer to the data in the constant buffer.
-	dataPtr = (MatrixBufferType*)mappedResource.pData;
-
-	// Copy the matrices into the constant buffer.
-	dataPtr->world = worldMatrix;
-	dataPtr->view = viewMatrix;
-	dataPtr->projection = projectionMatrix;
-
-	// Unlock the constant buffer.
-	deviceContext->Unmap(m_matrixBuffer, 0);
-
-	// Set the position of the constant buffer in the vertex shader.
-	bufferNumber = 0;
-
-	// Finally set the constant buffer in the vertex shader with the updated values.
-	deviceContext->VSSetConstantBuffers(bufferNumber, 1, &m_matrixBuffer);
-
-	return true;
-}
-
-
-void InstanceShaderClass::RenderShader(ID3D11DeviceContext* deviceContext, int indexCountPerInstance, int instanceCount)
-{
-	// Set the vertex input layout.
-	deviceContext->IASetInputLayout(m_layout);
-
-	// Set the vertex and pixel shaders that will be used to render this triangle.
-	deviceContext->VSSetShader(m_vertexShader, nullptr, 0);
-	deviceContext->PSSetShader(m_pixelShader, nullptr, 0);
-
 	// Render the geometry.
 	deviceContext->DrawIndexedInstanced(indexCountPerInstance, instanceCount, 0, 0, 0);
